fix: correctly resolve paths provided through Wrangler config on Windows (#13671)
* fix

fixes #13667

This PR uses the non-posix relative() method so that it correctly resolves when used with Windows paths that may be provided through the Wrangler configuration. It also converts the path to a POSIX path before embedding them into the worker script since backslashes get interpreted as escape characters and are lost in the process.

Also includes a minor fix where if the user had pages_build_output_dir set in their wrangler configuration, the adapter wasn't writing the worker to that directory (Cloudflare Pages expects the worker and assets to be in the same directory).